[exporterhelper] Move NewTracesRequest to exporterhelper/internal
