+++ exporter/exporterhelper/internal/new_request.go
@@ -11,6 +11,7 @@ import (
 	"go.opentelemetry.io/collector/exporter"
 	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/request"
 	"go.opentelemetry.io/collector/pdata/plog"
+	"go.opentelemetry.io/collector/pdata/ptrace"
 	"go.opentelemetry.io/collector/pipeline"
 	"go.uber.org/zap"
 )
@@ -61,6 +62,59 @@ func newConsumeLogs(converter request.RequestConverterFunc[plog.Logs], be *BaseE
 		if err != nil {
 			logger.Error("Failed to convert logs. Dropping data.",
 				zap.Int("dropped_log_records", ld.LogRecordCount()),
+				zap.Error(err))
+			return consumererror.NewPermanent(err)
+		}
+		return be.Send(ctx, req)
+	}
+}
+
+type tracesExporter struct {
+	*BaseExporter
+	consumer.Traces
+}
+
+// NewTracesRequest creates a new traces exporter based on a custom TracesConverter and Sender.
+// Experimental: This API is at the early stage of development and may change without backward compatibility
+// until https://github.com/open-telemetry/opentelemetry-collector/issues/8122 is resolved.
+func NewTracesRequest(
+	_ context.Context,
+	set exporter.Settings,
+	converter request.RequestConverterFunc[ptrace.Traces],
+	pusher request.RequestConsumeFunc,
+	options ...Option,
+) (exporter.Traces, error) {
+	if set.Logger == nil {
+		return nil, errNilLogger
+	}
+
+	if converter == nil {
+		return nil, errNilTracesConverter
+	}
+
+	if pusher == nil {
+		return nil, errNilConsumeRequest
+	}
+
+	be, err := NewBaseExporter(set, pipeline.SignalTraces, pusher, options...)
+	if err != nil {
+		return nil, err
+	}
+
+	tc, err := consumer.NewTraces(newConsumeTraces(converter, be, set.Logger), be.ConsumerOptions...)
+	if err != nil {
+		return nil, err
+	}
+
+	return &tracesExporter{BaseExporter: be, Traces: tc}, nil
+}
+
+func newConsumeTraces(converter request.RequestConverterFunc[ptrace.Traces], be *BaseExporter, logger *zap.Logger) consumer.ConsumeTracesFunc {
+	return func(ctx context.Context, td ptrace.Traces) error {
+		req, err := converter(ctx, td)
+		if err != nil {
+			logger.Error("Failed to convert traces. Dropping data.",
+				zap.Int("dropped_spans", td.SpanCount()),
 				zap.Error(err))
 			return consumererror.NewPermanent(err)
 		}
