+++ exporter/exporterhelper/traces.go
@@ -7,8 +7,6 @@ import (
 	"context"
 	"errors"
 
-	"go.uber.org/zap"
-
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/consumer"
 	"go.opentelemetry.io/collector/consumer/consumererror"
@@ -20,7 +18,6 @@ import (
 	"go.opentelemetry.io/collector/pdata/ptrace"
 	"go.opentelemetry.io/collector/pdata/xpdata/pref"
 	pdatareq "go.opentelemetry.io/collector/pdata/xpdata/request"
-	"go.opentelemetry.io/collector/pipeline"
 )
 
 var (
@@ -130,10 +127,6 @@ func (req *tracesRequest) BytesSize() int {
 	return tracesMarshaler.TracesSize(req.td)
 }
 
-type tracesExporter struct {
-	*internal.BaseExporter
-	consumer.Traces
-}
 
 // NewTraces creates an exporter.Traces that records observability metrics and wraps every request with a Span.
 func NewTraces(
@@ -149,7 +142,7 @@ func NewTraces(
 	if pusher == nil {
 		return nil, errNilPushTraces
 	}
-	return NewTracesRequest(ctx, set, requestFromTraces(), requestConsumeFromTraces(pusher),
+	return internal.NewTracesRequest(ctx, set, requestFromTraces(), requestConsumeFromTraces(pusher),
 		append([]Option{internal.WithQueueBatchSettings(NewTracesQueueBatchSettings())}, options...)...)
 }
 
@@ -167,50 +160,3 @@ func requestFromTraces() RequestConverterFunc[ptrace.Traces] {
 	}
 }
 
-// NewTracesRequest creates a new traces exporter based on a custom TracesConverter and Sender.
-// Experimental: This API is at the early stage of development and may change without backward compatibility
-// until https://github.com/open-telemetry/opentelemetry-collector/issues/8122 is resolved.
-func NewTracesRequest(
-	_ context.Context,
-	set exporter.Settings,
-	converter RequestConverterFunc[ptrace.Traces],
-	pusher RequestConsumeFunc,
-	options ...Option,
-) (exporter.Traces, error) {
-	if set.Logger == nil {
-		return nil, errNilLogger
-	}
-
-	if converter == nil {
-		return nil, errNilTracesConverter
-	}
-
-	if pusher == nil {
-		return nil, errNilConsumeRequest
-	}
-
-	be, err := internal.NewBaseExporter(set, pipeline.SignalTraces, pusher, options...)
-	if err != nil {
-		return nil, err
-	}
-
-	tc, err := consumer.NewTraces(newConsumeTraces(converter, be, set.Logger), be.ConsumerOptions...)
-	if err != nil {
-		return nil, err
-	}
-
-	return &tracesExporter{BaseExporter: be, Traces: tc}, nil
-}
-
-func newConsumeTraces(converter RequestConverterFunc[ptrace.Traces], be *internal.BaseExporter, logger *zap.Logger) consumer.ConsumeTracesFunc {
-	return func(ctx context.Context, td ptrace.Traces) error {
-		req, err := converter(ctx, td)
-		if err != nil {
-			logger.Error("Failed to convert traces. Dropping data.",
-				zap.Int("dropped_spans", td.SpanCount()),
-				zap.Error(err))
-			return consumererror.NewPermanent(err)
-		}
-		return be.Send(ctx, req)
-	}
-}
